Add casts to prevent integer overflow
extraBytesToRead is an unsigned short that is being subtracted from a
pointer.  When adding 1 to extraBytesToRead, extraBytesToRead could
potentially overflow before being subtracted from the pointer.
Casting extraBytesToRead to ptrdiff_t ensures extraBytesToRead is the
same size of a pointer, ensuring there is no loss of data before the
subtraction occurs.